Pull out android utils
