+++ packages/react-native/examples/react-native-cli/package.json
@@ -10,7 +10,7 @@
     "test": "jest"
   },
   "dependencies": {
-    "@honeybadger-io/react-native": "file:../../honeybadger-io-react-native-5.1.4_216.tgz",
+    "@honeybadger-io/react-native": "file:../../honeybadger-io-react-native-5.1.6_281.tgz",
     "react": "18.2.0",
     "react-native": "0.71.3"
   },
